Fix: Refatorando codigo, juntando funcoes de teste no mesmo arquivo.
